requirements.txt
patchright==1.55.2
python-dotenv==1.0.1
